post

+++ app/Http/Controllers/TasksController.php
@@ -10,20 +10,28 @@ class TasksController extends Controller
 {
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // タスク一覧ビューでそれを表示
+        return view('tasks.index',$data);
     }
 
     public function create()
     {
         $task = new Task;
 
-        // メッセージ作成ビューを表示
+        // タスク作成ビューを表示
         return view('tasks.create', [
             'task' => $task,
         ]);
@@ -37,10 +45,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
 
-        // メッセージを作成
+        // タスクを作成
         $task = new Task;
+        $task->user_id = \Auth::id(); 
         $task->content = $request->content;
-        $task->status = $request->status;
+        $task->status  = $request->status;
         $task->save();
 
         // トップページへリダイレクトさせる
@@ -49,24 +58,35 @@ class TasksController extends Controller
 
     public function show($id)
     {
-        // idの値でメッセージを検索して取得
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
 
-        // メッセージ詳細ビューでそれを表示
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+        // タスク詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
         ]);
+        }
+        
+        // トップページへリダイレクトさせる
+        return redirect('/');
     }
 
     public function edit($id)
     {
         // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
+        $task = \App\Task::findOrFail($id);
 
-        // メッセージ編集ビューでそれを表示
+        if (\Auth::id() === $task->user_id) {
+        // タスク編集ビューでそれを表示
         return view('tasks.edit', [
             'task' => $task,
         ]);
+        }
+        
+        // トップページへリダイレクトさせる
+        return redirect('/');
     }
 
     public function update(Request $request, $id)
@@ -77,23 +97,29 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを更新
+        // idの値でタスクを検索して取得
+        $task = \App\Task::findOrFail($id);
+        
+        // タスクを更新
+        if (\Auth::id() === $task->user_id) {
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
-
+        }
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
 
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値でタスクを検索して取得
+        $task = \App\Task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
